chore(clippy): fix transaction-dispatcher

+++ transaction-dispatcher/src/contracts/peg_manager.rs
@@ -320,8 +320,8 @@ pub(crate) fn decode_error(err: &alloy_contract::Error) -> Option<DomainErrors> 
             // Extract actual and expected values from the error
             // The InvalidSlotState error typically contains actual and expected slot states
             DomainErrors::InvalidSlotState {
-                expected: e.expected as u8,
-                actual: e.actual as u8,
+                expected: e.expected,
+                actual: e.actual,
             }
         }
         // Unhandled
